fix: high memory/cpu usage in dev
In development, the tutor-mfe plugin runs a separate container for every
MFE. This runs `npm run start` for each of the 9 MFEs, which causes the
developer laptops to crash and ice caps to melt.

Here, we adopt a different solution. By default, all MFEs run in a
single container, which is the same as the production container (with
caddy). Only the Caddyfile is different. This effectively resolves the
performance issue.

But to allow developers to hack on invidual MFEs, we also need to run
some of them in development mode, with `npm run start`. To do that, we
run separately only those MFEs for which the developer has a
bind-mounted directory that corresponds to the MFE repository. This
allows developers to keep working on MFEs with the help of `tutor mounts
...` command.